split pending action parsing and CTA confirmation tests
