+++ IntelligenceX.Chat/IntelligenceX.Chat.Tests/ChatServiceRoutingTrimTests.PendingActionsCtaConfirmation.cs
@@ -625,299 +625,4 @@ public sealed partial class ChatServiceRoutingTrimTests {
         Assert.False(selection.GetProperty("mutating").GetBoolean());
     }
 
-    [Fact]
-    public void ExpandContinuationUserRequest_DoesNotResolveSinglePendingActionWhenSingleOverlapIsNonTrailing() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            [Action]
-            ix:action:v1
-            id: act_failed4625
-            title: Run failed logon report (4625)
-            request: Run failed logon report on ADO Security and summarize the top five events.
-            mutating: false
-            reply: /act act_failed4625
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "run later" });
-        var expanded = Assert.IsType<string>(result);
-
-        Assert.Equal("run later", expanded);
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_ResolvesBestPendingActionByIntentOverlapWhenMultipleActionsExist() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            [Action]
-            ix:action:v1
-            id: act_failed4625
-            title: Failed logons (4625)
-            request: Run failed logon report on ADO Security.
-            mutating: false
-            reply: /act act_failed4625
-
-            [Action]
-            ix:action:v1
-            id: act_lockout4740
-            title: Account lockouts (4740)
-            request: Run account lockout report on ADO Security.
-            mutating: false
-            reply: /act act_lockout4740
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "failed logons please" });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        Assert.Equal("act_failed4625", doc.RootElement.GetProperty("ix_action_selection").GetProperty("id").GetString());
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_DoesNotResolveMultiplePendingActionsOnAmbiguousSingleTokenOverlap() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            [Action]
-            ix:action:v1
-            id: act_failed4625
-            title: Failed logons (4625)
-            request: Run failed logon report on ADO Security.
-            mutating: false
-            reply: /act act_failed4625
-
-            [Action]
-            ix:action:v1
-            id: act_lockout4740
-            title: Account lockouts (4740)
-            request: Run account lockout report on ADO Security.
-            mutating: false
-            reply: /act act_lockout4740
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "run it" });
-        var expanded = Assert.IsType<string>(result);
-
-        Assert.Equal("run it", expanded);
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_ResolvesFallbackBulletChoiceWithoutActionMarker() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            Absolutely — want me to pull top 5 recent events, or go straight to a focused cut like:
-            - Failed logons (4625)
-            - Account lockouts (4740)
-            - Logon activity mix (4624/4625/4634/4647)
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "failed logons please" });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        var selection = doc.RootElement.GetProperty("ix_action_selection");
-        Assert.Equal("choice_001", selection.GetProperty("id").GetString());
-        Assert.Equal("Failed logons (4625)", selection.GetProperty("request").GetString());
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_DoesNotResolveFallbackBulletChoiceWithoutPromptContext() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            Options
-            - Failed logons (4625)
-            - Account lockouts (4740)
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "failed logons please" });
-        var expanded = Assert.IsType<string>(result);
-
-        Assert.Equal("failed logons please", expanded);
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_ResolvesSingleNumberedFallbackChoiceWithPromptContext() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            Pulling failed logons from ADO now would be the next step, but I need your go-ahead in this flow:
-            1. Run failed logon report (4625) on ADO Security
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "go ahead and run it" });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        var selection = doc.RootElement.GetProperty("ix_action_selection");
-        Assert.Equal("choice_001", selection.GetProperty("id").GetString());
-        Assert.Equal("Run failed logon report (4625) on ADO Security", selection.GetProperty("request").GetString());
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_DoesNotResolveSingleBulletFallbackChoiceWithPromptContext() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            Pulling failed logons from ADO now would be the next step, but I need your go-ahead in this flow:
-            - Run failed logon report (4625) on ADO Security
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "go ahead and run it" });
-        var expanded = Assert.IsType<string>(result);
-
-        Assert.Equal("go ahead and run it", expanded);
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_ResolvesSingleBulletFallbackChoiceWithInlineActIdWithPromptContext() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            Pulling failed logons from ADO now would be the next step, but I need your go-ahead in this flow:
-            - Run failed logon report (4625) on ADO Security (/act act_failed4625)
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "go ahead and run it" });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        var selection = doc.RootElement.GetProperty("ix_action_selection");
-        Assert.Equal("act_failed4625", selection.GetProperty("id").GetString());
-        Assert.Equal("Run failed logon report (4625) on ADO Security", selection.GetProperty("request").GetString());
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_ResolvesSingleNumberedFallbackChoiceWithInlineActIdInParentheses() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            Pulling failed logons from ADO now would be the next step, but I need your go-ahead in this flow:
-            You can reply with /act act_failed4625 (or just 1) and I'll execute it immediately.
-            You can run one of these follow-up actions:
-            1. Run failed logon report (4625) on ADO Security (/act act_failed4625)
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "go ahead and run it" });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        var selection = doc.RootElement.GetProperty("ix_action_selection");
-        Assert.Equal("act_failed4625", selection.GetProperty("id").GetString());
-        Assert.Equal("Run failed logon report (4625) on ADO Security", selection.GetProperty("request").GetString());
-    }
-
-    [Fact]
-    public void ExpandContinuationUserRequest_ResolvesExplicitActForFallbackChoiceWithInlineActIdInParentheses() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            Pulling failed logons from ADO now would be the next step, but I need your go-ahead in this flow:
-            You can run one of these follow-up actions:
-            1. Run failed logon report (4625) on ADO Security (/act act_failed4625)
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "/act act_failed4625" });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        var selection = doc.RootElement.GetProperty("ix_action_selection");
-        Assert.Equal("act_failed4625", selection.GetProperty("id").GetString());
-        Assert.Equal("Run failed logon report (4625) on ADO Security", selection.GetProperty("request").GetString());
-    }
-
-    [Theory]
-    [InlineData("`/act act_001`")]
-    [InlineData("\"/act act_001\"")]
-    [InlineData("(/act act_001)")]
-    [InlineData("/act act_001.")]
-    [InlineData("/act act_001!")]
-    [InlineData("/act act_001?")]
-    public void ExpandContinuationUserRequest_ResolvesExplicitActWhenCommandUsesSafeWrappersOrPunctuation(string input) {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            [Action]
-            ix:action:v1
-            id: act_001
-            title: Failed logons (4625)
-            request: Run failed logons report.
-            mutating: false
-            reply: /act act_001
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", input });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        Assert.Equal("act_001", doc.RootElement.GetProperty("ix_action_selection").GetProperty("id").GetString());
-    }
-
-    [Theory]
-    [InlineData("/act act_001 please")]
-    [InlineData("/act act_001 and summarize")]
-    public void ExpandContinuationUserRequest_DoesNotResolveExplicitActWhenCommandHasTrailingText(string input) {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var assistantDraft = """
-            [Action]
-            ix:action:v1
-            id: act_001
-            title: Failed logons (4625)
-            request: Run failed logons report.
-            mutating: false
-            reply: /act act_001
-            """;
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", assistantDraft });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", input });
-        var expanded = Assert.IsType<string>(result);
-
-        Assert.Equal(input, expanded);
-    }
-
-    [Fact]
-    public void RememberPendingActions_NoMarkerOrFallbackChoices_DoesNotClearExistingActionContext() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var actionDraft = """
-            [Action]
-            ix:action:v1
-            id: act_001
-            title: Failed logons (4625)
-            request: Run failed logons report.
-            mutating: false
-            reply: /act act_001
-            """;
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", actionDraft });
-
-        var plainAssistantMessage = """
-            Quick read
-            normal service state flips, no spicy errors in this 5-event window.
-            """;
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", plainAssistantMessage });
-
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "/act act_001" });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        Assert.Equal("act_001", doc.RootElement.GetProperty("ix_action_selection").GetProperty("id").GetString());
-    }
-
-    [Fact]
-    public void RememberPendingActions_RehydratesReplayActionFromExecutionContractBlockerText() {
-        var session = new ChatServiceSession(new ServiceOptions(), Stream.Null);
-        var selectionPayload = "{\"ix_action_selection\":{\"id\":\"act_failed4625\",\"title\":\"Failed logons (4625)\",\"request\":\"Run failed logon report on ADO Security and summarize top events.\",\"mutating\":false}}";
-        var blockerText = Assert.IsType<string>(BuildExecutionContractBlockerTextMethod.Invoke(
-            null,
-            new object?[] { selectionPayload, "Ok, doing it now.", "no_tool_calls_after_watchdog_retry" }));
-
-        RememberPendingActionsMethod.Invoke(session, new object?[] { "thread-001", blockerText });
-        var result = ExpandContinuationUserRequestMethod.Invoke(session, new object?[] { "thread-001", "go ahead and run it" });
-        var expanded = Assert.IsType<string>(result);
-
-        using var doc = JsonDocument.Parse(expanded);
-        var selection = doc.RootElement.GetProperty("ix_action_selection");
-        Assert.Equal("act_failed4625", selection.GetProperty("id").GetString());
-    }
 }
